1.3 temperature with float

+++ exercises/temperature.c
@@ -2,17 +2,17 @@
 
 main() {
 
-	int fahr, celsius;
+	float fahr, celsius;
 	int lower, upper, step;
 
 	lower = 0;
 	upper = 300;
 	step = 20;
 	
-	fahr = lower;
+	fahr = (float) lower;
 	while (fahr <= upper) {
-		celsius = 5 * (fahr-32) / 9;
-		printf("%3d %3d\n", fahr, celsius);
+		celsius = 5.0 * (fahr-32.0) / 9.0;
+		printf("%3.0f %6.1f\n", fahr, celsius);
 		fahr = fahr + step;
 	}
 }
